refactor(rust cycle 18): cr_handlers/ split by HTTP verb (≤300 LOC)
