+++ src/bin/email_api_dir/admin_ops/cr_handlers/delete.rs
@@ -1,0 +1,25 @@
+#![allow(unused_imports, dead_code)]
+use super::super::*;
+
+pub(crate) async fn api_admin_change_request_delete(
+    path: web::Path<String>,
+    mongo: web::Data<Arc<mongodb::Client>>,
+) -> impl Responder {
+    let id = path.into_inner();
+    let coll = mongo
+        .database(&mongo_db_name())
+        .collection::<ChangeRequestItem>(ADMIN_CHANGE_REQUESTS_COLL);
+
+    match coll.delete_one(doc! { "id": &id }).await {
+        Ok(res) if res.deleted_count > 0 => {
+            HttpResponse::Ok().json(serde_json::json!({ "deleted": true, "id": id }))
+        }
+        Ok(_) => HttpResponse::NotFound()
+            .json(serde_json::json!({ "deleted": false, "message": "Change request not found" })),
+        Err(e) => {
+            eprintln!("api_admin_change_request_delete error: {}", e);
+            HttpResponse::InternalServerError()
+                .json(serde_json::json!({ "message": "Failed to delete change request" }))
+        }
+    }
+}
